+++ crates/types/src/chain_parameters.rs
@@ -12,11 +12,11 @@ impl ChainParameters {
         ChainParameters { chain_id: 1, base_fee_params: BaseFeeParams::ethereum() }
     }
 
-    pub fn calc_next_block_base_fee(&self, gas_used: u128, gas_limit: u128, base_fee: u128) -> u128 {
+    pub fn calc_next_block_base_fee(&self, gas_used: u64, gas_limit: u64, base_fee: u64) -> u64 {
         self.base_fee_params.next_block_base_fee(gas_used, gas_limit, base_fee)
     }
 
-    pub fn calc_next_block_base_fee_from_header(&self, header: &Header) -> u128 {
+    pub fn calc_next_block_base_fee_from_header(&self, header: &Header) -> u64 {
         self.base_fee_params.next_block_base_fee(header.gas_used, header.gas_limit, header.base_fee_per_gas.unwrap_or_default())
     }
 }
